Ooo, a few updates, main one being, disabling google assistant
that's right, it had a vulnerability and it also refused to install using "npm i" or "npm i google-assistant". there is an alternative, "npm i nodejs-assistant", but it also fails to install. both fail to install due to node-gyp-blahblahblah failing to work. i forgot it's full name, but you can look it up, its under the grpc library. oh and i added the dev packages and will need to add tests, for now my mission is to containerize this under docker :smirk_cat: :smiley: